fix flash
